Enhance project structure and file organization
