+++ .github/CODE_OF_CONDUCT.md
@@ -1,9 +1,9 @@
 # Contributor Covenant Code of Conduct
 
-## 📌 Our Commitment
+## Our Commitment
 We are committed to fostering a welcoming, inclusive, and respectful environment for all contributors. We encourage open collaboration regardless of experience level, gender identity, race, nationality, or background.
 
-## 🤝 Our Standards
+## Our Standards
 We expect all contributors to:
 - Be **respectful and considerate** in all interactions.
 - Provide **constructive feedback** and help others grow.
@@ -17,20 +17,20 @@ Unacceptable behavior includes:
 - Publishing private information without consent.
 - Any other behavior deemed inappropriate in a professional setting.
 
-## 🔍 Reporting Issues
+## Reporting Issues
 If you witness or experience any violations of this Code of Conduct:
 - Contact the project maintainer at **ameerhamza92099@gmail.com**.
 - Alternatively, report the issue by opening a confidential issue on GitHub.
 
 All reports will be handled **confidentially** and with fairness.
 
-## ⚖️ Enforcement
+## Enforcement
 Any violations may result in:
 - A **formal warning** from the maintainers.
 - Temporary or permanent **ban from contributing**.
 - Other actions deemed necessary by the project team.
 
-## 🏆 Acknowledgment
+## Acknowledgment
 This Code of Conduct is based on the [Contributor Covenant](https://www.contributor-covenant.org/) v2.1.
 
 By contributing to this project, you agree to abide by this Code of Conduct.
