+++ Library Mnagment System/src/Book.cs
@@ -1,0 +1,7 @@
+class Book
+{
+    public int Id { get; set; }
+    public string Title { get; set; }
+    public string Author { get; set; }
+    public bool IsBorrowed { get; set; } = false;
+}
