+++ Library Mnagment System/src/Library.cs
@@ -1,13 +1,5 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
-
-class Book
-{
-    public int Id { get; set; }
-    public string Title { get; set; }
-    public string Author { get; set; }
-    public bool IsBorrowed { get; set; } = false;
-}
 
 class Library
 {
@@ -139,66 +131,6 @@ class Library
         catch (Exception ex)
         {
             Console.WriteLine($"Error: {ex.Message}\n");
-        }
-    }
-}
-
-class Program
-{
-    static void Main()
-    {
-        Library library = new Library();
-        while (true)
-        {
-            Console.WriteLine("\n========================");
-            Console.WriteLine(" Library Management System ");
-            Console.WriteLine("========================");
-            Console.WriteLine("1. Add Book");
-            Console.WriteLine("2. View Books");
-            Console.WriteLine("3. Borrow Book");
-            Console.WriteLine("4. Return Book");
-            Console.WriteLine("5. Remove Book");
-            Console.WriteLine("6. Exit");
-            Console.Write("Enter your choice: ");
-
-            try
-            {
-                if (int.TryParse(Console.ReadLine(), out int choice))
-                {
-                    switch (choice)
-                    {
-                        case 1:
-                            library.AddBook();
-                            break;
-                        case 2:
-                            library.ViewBooks();
-                            break;
-                        case 3:
-                            library.BorrowBook();
-                            break;
-                        case 4:
-                            library.ReturnBook();
-                            break;
-                        case 5:
-                            library.RemoveBook();
-                            break;
-                        case 6:
-                            Console.WriteLine("\nExiting the program...\n");
-                            return;
-                        default:
-                            Console.WriteLine("\nError: Invalid choice! Please select a valid option.\n");
-                            break;
-                    }
-                }
-                else
-                {
-                    Console.WriteLine("\nError: Invalid input! Please enter a number.\n");
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error: {ex.Message}\n");
-            }
         }
     }
 }
