+++ README.md
@@ -1,38 +1,41 @@
 # Library Management System (Console App)
 
-## 📌 Overview
+## Overview
 The **Library Management System** is a simple console-based application built using **C#** that allows users to manage a collection of books efficiently. It provides functionalities to add, view, borrow, return, and remove books from the library. This project is designed for learning purposes and can be extended to include more advanced features.
 
-## 🚀 Features
-- 📚 **Add Book**: Add new books with a title and author.
-- 👀 **View Books**: List all available books with their status.
-- 🔄 **Borrow Book**: Mark a book as borrowed.
-- 🔄 **Return Book**: Mark a book as returned.
-- ❌ **Remove Book**: Delete a book from the library.
-- 🎭 **User-Friendly Interface**: Simple and easy-to-use console-based interface.
+## Features
+- **Add Book**: Add new books with a title and author.
+- **View Books**: List all available books with their status.
+- **Borrow Book**: Mark a book as borrowed.
+- **Return Book**: Mark a book as returned.
+- **Remove Book**: Delete a book from the library.
+- **User-Friendly Interface**: Simple and easy-to-use console-based interface.
 
-## 🛠️ Technologies Used
+## Technologies Used
 - **Programming Language**: C#
 - **Development Environment**: .NET Console Application
 
-## 📂 Project Structure
+## Project Structure
 ```
-📦 Library Management System
-├── 📄 Library Management System.sln  # Solution file
-├── 📂 Library Management System      # Main project folder
-│   ├── 📄 Program.cs                 # Main application file
-├── 📂 .github                        # GitHub-related files
-│   ├── 📄 ISSUE_TEMPLATE             # Issue template
-├── 📄 .gitattributes                 # Git attributes file
-├── 📄 .gitignore                     # Git ignore file
-├── 📄 CODE_OF_CONDUCT.md              # Code of Conduct
-├── 📄 CONTRIBUTING.md                 # Contribution guidelines
-├── 📄 LICENSE                         # MIT License
-├── 📄 README.md                       # Project documentation
-├── 📄 SECURITY.md                      # Security policies
+Library Management System
+├── Library Management System.sln  # Solution file
+├── Library Management System      # Main project folder
+│   ├── src                        # Source folder
+│   │   ├── Program.cs             # Main application file
+│   │   ├── Book.cs                # Book class file
+│   │   ├── Library.cs             # Library class file
+├── .github                        # GitHub-related files
+│   ├── ISSUE_TEMPLATE             # Issue template
+│   ├── CODE_OF_CONDUCT.md         # Code of Conduct
+│   ├── CONTRIBUTING.md            # Contribution guidelines
+├── .gitattributes                 # Git attributes file
+├── .gitignore                     # Git ignore file
+├── LICENSE                        # MIT License
+├── README.md                      # Project documentation
+├── SECURITY.md                    # Security policies
 ```
 
-## 🏗️ Installation & Usage
+## Installation & Usage
 
 ### Prerequisites
 - Install **.NET SDK** from [Microsoft .NET](https://dotnet.microsoft.com/en-us/download)
@@ -42,16 +45,20 @@ The **Library Management System** is a simple console-based application built us
    ```sh
    git clone https://github.com/Hamza-92/Library-Management-System-Console-App.git
    ```
-1. **Navigate to the project directory**
+2. **Navigate to the project directory**
    ```sh
    cd "Library Management System"
    ```
-3. **Run the Application**
+3. **Restore Dependencies**
+   ```sh
+   dotnet restore
+   ```
+4. **Run the Application**
    ```sh
    dotnet run
    ```
 
-## 📌 How to Use
+## How to Use
 1. Run the application.
 2. Choose an option from the menu:
    - `1` to Add a Book
@@ -61,16 +68,15 @@ The **Library Management System** is a simple console-based application built us
    - `5` to Remove a Book
    - `6` to Exit
 
-## 🤝 Contributing
+## Contributing
 We welcome contributions! Please read our [CONTRIBUTING.md](CONTRIBUTING.md) to get started.
 
-## 🛡️ Security Policy
+## Security Policy
 For reporting vulnerabilities, please refer to our [SECURITY.md](SECURITY.md).
 
-## 📝 License
+## License
 This project is licensed under the **MIT License** – see the [LICENSE](LICENSE) file for details.
 
-## 📬 Contact
+## Contact
 - **GitHub**: [Hamza-92](https://github.com/Hamza-92)
 - **Email**: ameerhamza92099@gmail.com
-
